requirements.txt
requests
pandas
pillow
pytesseract